pyproject.toml
[project]
name = "django-notion-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django>=5.1.3",
  "djangorestframework>=3.15.2",
  "djangoba-tools",
  "django-environ>=0.11.2",
  "django-health-check>=3.18.3",
  "django-silk>=5.3.1",
  "drf-yasg>=1.21.8",
  "djangorestframework-simplejwt[crypto]>=5.3.1",
  "djoser>=2.3.1",
  "django-allow-cidr>=0.7.1",
  "django-filter>=24.3",
  "python-dotenv>=1.0.1",
  "psycopg>=3.2.3",
  "whitenoise>=6.8.2",
  "django-cors-headers>=4.6.0",
  "django-extensions>=3.2.3",
]

[tool.uv.sources]
djangoba-tools = { path = "../djangoba-tools" }

[dependency-groups]
dev = ["ipython>=8.29.0", "ruff>=0.8.0"]
